PUT /missions/ updating routeDetails fixed

+++ back-end/src/main/scala/model/RouteDetails.scala
@@ -31,7 +31,7 @@ case class RouteDetails private (
     ("id" -> id) ~
       ("start" -> start) ~
       ("end" -> end) ~
-      ("distance" -> distance) ~
+      ("distance" -> distance.toInt) ~
       ("points" -> parse(points)) ~
       ("noneNormalSegments" -> parse(noneNormalSegments))
   }
